+++ src/pas/plugins/authomatic/browser/controlpanel.py
@@ -1,6 +1,10 @@
 from pas.plugins.authomatic.interfaces import _
+from pas.plugins.authomatic.interfaces import IPasPluginsAuthomaticLayer
 from pas.plugins.authomatic.interfaces import IPasPluginsAuthomaticSettings
 from plone.app.registry.browser import controlpanel
+from plone.restapi.controlpanels import RegistryConfigletPanel
+from zope.component import adapter
+from zope.interface import Interface
 
 
 class AuthomaticSettingsEditForm(controlpanel.RegistryEditForm):
@@ -21,3 +25,15 @@ class AuthomaticSettingsEditFormSettingsControlPanel(
     controlpanel.ControlPanelFormWrapper
 ):
     form = AuthomaticSettingsEditForm
+
+
+@adapter(Interface, IPasPluginsAuthomaticLayer)
+class AuthomaticSettingsConfigletPanel(RegistryConfigletPanel):
+    """Control Panel endpoint"""
+
+    schema = IPasPluginsAuthomaticSettings
+    configlet_id = "authomatic"
+    configlet_category_id = "Products"
+    title = _("Authomatic settings")
+    group = ""
+    schema_prefix = "pas.plugins.authomatic.interfaces.IPasPluginsAuthomaticSettings"